Feat/proc gen

+++ Client/src/main/java/com/jagex/map/SceneTileData.java
@@ -10,7 +10,7 @@ import com.rspsi.misc.Location;
 
 public class SceneTileData {
 
-	private byte overlayId = -1, underlayId = -1;
+	private short overlayId = -1, underlayId = -1;
 	private byte overlayOrientation = -1, overlayType = -1;
 
 	private int tileHeight = -1;
@@ -100,7 +100,7 @@ public class SceneTileData {
 				groundDecoConfig = tile.groundDecoration.getConfig();
 			}
 	}
-	
+
 	public SceneTileData(CopyOptions copyWindow, SceneTile tile, int tileX, int tileY) {
 		if (copyWindow.copyGameObjects())
 			if (tile.objectCount > 0) {
@@ -152,7 +152,7 @@ public class SceneTileData {
 		return groundDecoId;
 	}
 
-	public byte getOverlayId() {
+	public short getOverlayId() {
 		return overlayId;
 	}
 
@@ -172,7 +172,7 @@ public class SceneTileData {
 		return tileHeight;
 	}
 
-	public byte getUnderlayId() {
+	public short getUnderlayId() {
 		return underlayId;
 	}
 
@@ -220,7 +220,7 @@ public class SceneTileData {
 		this.groundDecoId = groundDecoId;
 	}
 
-	public void setOverlayId(byte overlayId) {
+	public void setOverlayId(short overlayId) {
 		this.overlayId = overlayId;
 	}
 
@@ -240,7 +240,7 @@ public class SceneTileData {
 		this.tileHeight = tileHeight;
 	}
 
-	public void setUnderlayId(byte underlayId) {
+	public void setUnderlayId(short underlayId) {
 		this.underlayId = underlayId;
 	}
