+++ Client/src/main/java/com/rspsi/game/save/tile/state/OverlayState.java
@@ -7,16 +7,16 @@ import lombok.EqualsAndHashCode;
 @Data
 @EqualsAndHashCode(callSuper=false)
 public class OverlayState extends TileState {
-	
-	private byte id;
+
+	private short id;
 	private byte rotation;
 	private byte shape;
-	
+
 	public OverlayState(int x, int y, int z) {
 		super(x, y, z);
 	}
 
-	public byte getId() {
+	public short getId() {
 		return id;
 	}
 	public byte getRotation() {
@@ -25,7 +25,7 @@ public class OverlayState extends TileState {
 	public byte getShape() {
 		return shape;
 	}
-	
+
 	@Override
 	public void preserve() {
 		this.id = Client.getSingleton().mapRegion.overlays[z][x][y];
@@ -37,7 +37,7 @@ public class OverlayState extends TileState {
 	public int getUniqueId() {
 		return 2;
 	}
-	
-	
+
+
 
 }
